Changed ripl.py to return more information for assume,predict,observe.
It used to return only the value for assume and predict, and nothing
for observe. Now it returns whatever ripl.execute_instruction
returns, which is currently a dict.

Also changed ripl.list_directives() to return a value along with each
dict. This may eventually be refactored so that a lower layer always
returns values with directives.